Rem UsageParams, add all props to ApplianceUsage
I'm thinking it doesn't make sense to have a separate value object for
UsageParams. For example, what if I just want to change the quantity?
I'll have to construct a whole new UsageParams object.

Also this way I can try to apply restrictions on the NumHours based on the
total number of hours that solar is available at the Site.