Specify Site FK to make it parent
Gets rid of this error when migrating -

System.AggregateException : One or more errors occurred. (The child/dependent side could not be determined for the one-to-one relationship between 'Consumption.Site' and 'Site.Consumption'. To identify the child/dependent side of the relationship, configure the foreign key property. If these navigations should not be part of the same relationship configure them without specifying the inverse. See http://go.microsoft.com/fwlink/?LinkId=724062 for more details.)