Use Secret Manager to add postgres secrets
Just adding password for now. Secret Manager can be used like so:
```
# Initialize
dotnet user-secrets init

# Set
dotnet user-secrets set "PostgresPassword" "SomeSuperStrongPassword"
```
This will add a UserSecretsIdAttribute to the project. The contents of the attribute can be anything but need to be unique inside the project.
The Secret Manager will auto-populate the contents with a Guid but I replace the Guid with my own meaningful name.
When I do this I have to make sure to also change the name of the generated secret folder containing the secrets.json to the name assigned by me.
This folder is under ~/.microsoft/usersecrets for macOS and probably %APPDATA%\Microsoft\UserSecrets for Windows.