Ren table to UsageIntervals, reintro the manual Id setting
I did this cuz apparently sqlite could not find the NOT NULL constraint on the table of UsageTimeIntervals.

So first I tried renaming the table, then when that didn't work I bit the bullet and restored to the manual Id setting.

Whatever it works now, don't wanna think about it.